requirements.txt
django
djangorestframework
mysqlclient
python-dotenv
djangorestframework-simplejwt